* moving animation config
