+++ TypeScript/Edt-Sledt/src/automation/animation.ts
@@ -1,8 +1,8 @@
 import { Actions, Actions$ } from '../../../Shared/actions/actions';
 import { filter, map } from 'rxjs/operators';
 import { midiCCAutomation$ } from '../communication/midi';
-import { AnimationTypes } from '../../../Shared/vidt-presets';
 import { animationTypeCCNumber, automationChannel } from '../../config/config';
+import { AnimationTypes } from '../../../Shared/vidt/animation';
 
 export const animationTypeCC$ = Actions$.animationType.pipe(
     map(type => ({
