+++ TypeScript/Shared/vidt/animation.ts
@@ -1,0 +1,12 @@
+import { enumToArray } from '../utils/utils';
+
+export enum AnimationTypes {
+    '____EMPTY____',
+    bounce,
+    rotate,
+    spin,
+    stretch,
+    mirror,
+}
+
+export const animationTypeArr = enumToArray(AnimationTypes);
